1.d partially complete.
Grid searching ElasticNet takes a really long time, probably because there
are three parameters to cycle through. It throws some errors about alpha=0,
but appears to come up with good results.

The plots of score dependence on alpha for Ridge are okay, could probably
copy and paste that code for the other estimators and get okay plots.